split MINIT for Category
