+++ src/php/classes/category.stub.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * Copyright (c) php-ecma-intl contributors.
+ *
+ * This source file is subject to the BSD-3-Clause license that is bundled with
+ * this package in the file LICENSE and is available at the following web
+ * address: https://opensource.org/license/bsd-3-clause/
+ *
+ * This source file may utilize copyrighted material from third-party open
+ * source projects, the use of which is acknowledged in the NOTICE file bundled
+ * with this package.
+ *
+ * @generate-class-entries
+ */
+
+namespace Ecma\Intl
+{
+    /**
+     * Categories of values supported by this implementation.
+     *
+     * @link https://tc39.es/ecma402/#sec-intl.supportedvaluesof ECMA-402, section 8.3.2, Intl.supportedValuesOf
+     */
+    enum Category: string
+    {
+        case Calendar = 'calendar';
+        case Collation = 'collation';
+        case Currency = 'currency';
+        case NumberingSystem = 'numberingSystem';
+        case TimeZone = 'timeZone';
+        case Unit = 'unit';
+    }
+}
