+++ src/php/classes/category_arginfo.h
@@ -1,0 +1,46 @@
+/* This is a generated file, edit the .stub.php file instead.
+ * Stub hash: 430525d27dcfa691b8f1e7204cbc53adb2db059f */
+
+
+
+
+static const zend_function_entry class_Ecma_Intl_Category_methods[] = {
+	ZEND_FE_END
+};
+
+static zend_class_entry *register_class_Ecma_Intl_Category(void)
+{
+	zend_class_entry *class_entry = zend_register_internal_enum("Ecma\\Intl\\Category", IS_STRING, class_Ecma_Intl_Category_methods);
+
+	zval enum_case_Calendar_value;
+	zend_string *enum_case_Calendar_value_str = zend_string_init("calendar", strlen("calendar"), 1);
+	ZVAL_STR(&enum_case_Calendar_value, enum_case_Calendar_value_str);
+	zend_enum_add_case_cstr(class_entry, "Calendar", &enum_case_Calendar_value);
+
+	zval enum_case_Collation_value;
+	zend_string *enum_case_Collation_value_str = zend_string_init("collation", strlen("collation"), 1);
+	ZVAL_STR(&enum_case_Collation_value, enum_case_Collation_value_str);
+	zend_enum_add_case_cstr(class_entry, "Collation", &enum_case_Collation_value);
+
+	zval enum_case_Currency_value;
+	zend_string *enum_case_Currency_value_str = zend_string_init("currency", strlen("currency"), 1);
+	ZVAL_STR(&enum_case_Currency_value, enum_case_Currency_value_str);
+	zend_enum_add_case_cstr(class_entry, "Currency", &enum_case_Currency_value);
+
+	zval enum_case_NumberingSystem_value;
+	zend_string *enum_case_NumberingSystem_value_str = zend_string_init("numberingSystem", strlen("numberingSystem"), 1);
+	ZVAL_STR(&enum_case_NumberingSystem_value, enum_case_NumberingSystem_value_str);
+	zend_enum_add_case_cstr(class_entry, "NumberingSystem", &enum_case_NumberingSystem_value);
+
+	zval enum_case_TimeZone_value;
+	zend_string *enum_case_TimeZone_value_str = zend_string_init("timeZone", strlen("timeZone"), 1);
+	ZVAL_STR(&enum_case_TimeZone_value, enum_case_TimeZone_value_str);
+	zend_enum_add_case_cstr(class_entry, "TimeZone", &enum_case_TimeZone_value);
+
+	zval enum_case_Unit_value;
+	zend_string *enum_case_Unit_value_str = zend_string_init("unit", strlen("unit"), 1);
+	ZVAL_STR(&enum_case_Unit_value, enum_case_Unit_value_str);
+	zend_enum_add_case_cstr(class_entry, "Unit", &enum_case_Unit_value);
+
+	return class_entry;
+}
