+++ src/php/ecma_intl.c
@@ -48,7 +48,7 @@ ZEND_GET_MODULE(ecma_intl)
 
 PHP_MINIT_FUNCTION(ecma_intl) {
   registerEcmaIntl();
-  registerEcmaIntlCategory();
+  PHP_MINIT(ecma_intl_category)(INIT_FUNC_ARGS_PASSTHRU);
   PHP_MINIT(ecma_intl_locale)(INIT_FUNC_ARGS_PASSTHRU);
   PHP_MINIT(ecma_intl_locale_characterdirection)(INIT_FUNC_ARGS_PASSTHRU);
   PHP_MINIT(ecma_intl_locale_options)(INIT_FUNC_ARGS_PASSTHRU);
